Enemy attack and push
A script named EnemyHitbox has been created and connected to enemy. Now the enemy will push the player when attacking and take health from the player. The player has also been updated to push the enemy when attacking.

NOTE: The player is still unable to die when losing all health but the enemy can die.