Added Scene Change & Crossroad Sprites
Added a functioning Scene Change mechanic, which works at a low level. To change scenes add an object with a Box Collider 2D and the SceneChange script. Enter the name of the scene you want to change to and then when the player collides with the object the scene will change. The scene needs to be added to the active scenes in the build settings for this to work. I have also begun working on the Crossroad scene.