Implemented Scene Save/Load Functionality
I have added code to the GameManager script that allows the transitions between scenes to not cause the game to reset values that relate to the player and their progress, i.e. gold, etc. This also will allow the spawn point of the player in a scene to depend on the last scene that they were in.