Updated Player & Home Scene
I have updated the player prefab hitbox again and reverted it to be a smaller size than the full sprite. It may cause issues later on but for now it makes navigation in buildings feel better. Have fixed issues with the collision tileset in the Home scene. Have also fixed an issue where the camera was animating its movement to the player on loading a new scene.