Weapon script
Made minor changes in preperation for a new attack option. Tomorrow, I will code in a way to attack in the opposite direction you are looking. This will allow the player to attack while running rather than turn towards the enemy and then attack which usually results in taking damage.