Modified Main Menu & Added FTM to Scene
I have centralised the code for the main menu into one script as opposed to two. There are still issues with the Animations in the Main Menu but the functionality is working. I have added Floating Text Manager to the Home scene so the functionality that relies on it should be working now.