Added SceneChangeConditional
I have added a variation on the SceneChange script that allows a condition for allowing the player to change scenes. I have also changed many scripts associated with the food counter object in order to try and get the functionality for that working. It is not working as of this build.