Dungeon transition
Made it so the player can enter the dungeon only once per day. Also made it so the player can only enter the second dungeon after the gaurdian in the first dungeon is killed.

NOTE: In regards to only being able to go to the dungeon once per day, this means if you enter the first dungeon and then immediately exit it, you cannot enter again until the next in game day.